requirements.txt
numpy
pandas
scikit-learn
sdv
tensorflow
tensorflow_io
python-dotenv
django
matplotlib
seaborn
fastparquet